+++ aa_forum/__init__.py
@@ -5,5 +5,5 @@ AA-Forum
 # Django
 from django.utils.translation import gettext_lazy as _
 
-__version__ = "1.19.5"
+__version__ = "2.0.0-beta.1"
 __title__ = _("Forum")